+++ .github/release.yml
@@ -25,6 +25,6 @@ changelog:
     - title: CI
       labels: ['Type: CI']
     - title: Dependency Updates
-      labels: ['Type: Dependencies', "dependencies"]
+      labels: ['Type: Dependencies', dependencies]
     - title: Other Changes
       labels: ['*']
