+++ .github/workflows/release.yaml
@@ -6,7 +6,7 @@ env:
 on:
   push:
     tags:
-      - "*"
+      - '*'
 
 permissions:
   contents: write
