+++ .github/workflows/test.yaml
@@ -8,8 +8,8 @@ on:
       - main
     paths:
       - '**.ts'
-      - '.github/workflows/test.yml'
-      - 'deno.json'
+      - .github/workflows/test.yml
+      - deno.json
   pull_request:
     branches:
       - main
@@ -94,7 +94,7 @@ jobs:
       - run: deno publish --dry-run
 
   action-timeline:
-    needs: 
+    needs:
       - ci
       - test
       - coverage-report
@@ -102,4 +102,4 @@ jobs:
     if: ${{ always() }}
     runs-on: ubuntu-latest
     steps:
-    - uses: Kesin11/actions-timeline@427ee2cf860166e404d0d69b4f2b24012bb7af4f # v2.2.3
+      - uses: Kesin11/actions-timeline@427ee2cf860166e404d0d69b4f2b24012bb7af4f # v2.2.3
