+++ .gitignore
@@ -3,3 +3,4 @@
 /types
 /.coverage
 /npm
+/dist
