+++ .npmrc
@@ -1,0 +1,1 @@
+use-node-version=22.14.0
