+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "deno.enable": true
+	"deno.enable": true
 }
