+++ build.config.ts
@@ -1,0 +1,12 @@
+import { defineBuildConfig } from 'unbuild';
+
+export default defineBuildConfig({
+	entries: [
+		'src/index',
+	],
+	declaration: 'node16',
+	clean: true,
+	externals: [
+		'@typescript-eslint/utils',
+	],
+});
