--- deno.json
@@ -1,35 +1,0 @@
-{
-  "name": "@ryoppippi/eslint-plugin",
-  "version": "0.0.0",
-  "tasks": {
-    "dev": "deno run --watch mod.ts",
-    "check": "deno check ./**/*.ts && deno fmt --check",
-    "lint": "deno lint ./**/*.ts",
-    "test": "deno test -A --parallel --doc",
-    "test:coverage": "deno task test --coverage=.coverage",
-    "coverage:show": "deno coverage .coverage",
-    "coverage:lco": "deno coverage --lcov .coverage > .coverage/lcov.info"
-  },
-  "fmt": {
-    "exclude": [
-      "*.md"
-    ]
-  },
-  "imports": {
-    "@cross/test": "jsr:@cross/test@^0.0.9",
-    "@std/assert": "jsr:@std/assert@^1.0.0",
-    "@typescript-eslint/typescript-estree": "npm:@typescript-eslint/typescript-estree@^8.0.0",
-    "@typescript-eslint/utils": "npm:@typescript-eslint/utils@^8.0.0",
-    "eslint": "npm:eslint@^9.5.0"
-  },
-  "exports": "./mod.ts",
-  "publish": {
-    "exclude": [
-      ".vscode",
-      ".github",
-      ".gitignore",
-      ".coverage"
-    ]
-  },
-  "keywords": []
-}
