--- deps.ts
@@ -1,5 +1,0 @@
-import * as eslint from "eslint";
-
-export { eslint };
-export type ESLint = eslint.ESLint;
-export type Plugin = eslint.ESLint.Plugin;
