+++ eslint.config.js
@@ -1,0 +1,9 @@
+import { ryoppippi } from '@ryoppippi/eslint-config';
+
+export default ryoppippi({
+	type: 'lib', // or 'lib'
+	svelte: false,
+	typescript: {
+		tsconfigPath: './tsconfig.json',
+	},
+});
