+++ src/rules/require-comment-on-useEffect.md
@@ -11,7 +11,7 @@ const someValue = 1;
 
 // prints the value of `someValue` whenever it changes
 useEffect(() => {
-  console.log(someValue);
+	console.log(someValue);
 }, [someValue]);
 ```
 
@@ -20,6 +20,6 @@ useEffect(() => {
 
 const someValue = 1;
 useEffect(() => {
-  console.log(someValue);
+	console.log(someValue);
 }, [someValue]);
 ```
